requirements.txt
streamlit==1.32.0
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.1
langchain-core==0.3.0
langchain-groq==0.2.0
langgraph==0.2.16